passed initial test of Cnn_conv...... milestone

+++ RL/src/Activation.cpp
@@ -1,25 +1,25 @@
 #include "Activation.h"
 
-double LinearActivation::activate(const double x) {
+float LinearActivation::activate(const float x) {
   return x;
 }
 
-double LinearActivation::dActivate(const double x) {
+float LinearActivation::dActivate(const float x) {
   return 1;
 }
 
-double RELU::activate(const double x) {
+float RELU::activate(const float x) {
   return x > 0 ? x : 0.0;
 }
 
-double RELU::dActivate(const double x) {
+float RELU::dActivate(const float x) {
   return x > 0 ? 1.0 : 0.0;
 }
 
-double LeakyRELU::activate(const double x) {
+float LeakyRELU::activate(const float x) {
   return x > 0 ? x : (a * x);
 }
 
-double LeakyRELU::dActivate(const double x) {
+float LeakyRELU::dActivate(const float x) {
   return x > 0 ? 1.0 : a;
 }
